# fix: Test Infrastructure Phase 1 - Critical Issues Resolved