… docs/adr/ADR-017-article-vii-meta-constitutional-evolution.md
… plans/article-vii-phase1-telemetry.md